Split page files to component
